+++ packages/snaps-controllers/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- **BREAKING:** All action types were renamed from `DoSomething` to `ControllerNameDoSomethingAction` ([#3907](https://github.com/MetaMask/snaps/pull/3907), [#3911](https://github.com/MetaMask/snaps/pull/3911))
+- **BREAKING:** All action types were renamed from `DoSomething` to `ControllerNameDoSomethingAction` ([#3907](https://github.com/MetaMask/snaps/pull/3907), [#3911](https://github.com/MetaMask/snaps/pull/3911), [#3912](https://github.com/MetaMask/snaps/pull/3912))
   - `SnapController` actions:
     - `GetSnap` is now `SnapControllerGetSnapAction`.
       - Note: The method is now called `getSnap` instead of `get`.
@@ -44,6 +44,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `Schedule` is now `CronjobControllerScheduleAction`.
     - `Cancel` is now `CronjobControllerCancelAction`.
     - `Get` is now `CronjobControllerGetAction`.
+  - `SnapInterfaceController` actions:
+    - `CreateInterface` is now `SnapInterfaceControllerCreateInterfaceAction`.
+    - `GetInterface` is now `SnapInterfaceControllerGetInterfaceAction`.
+    - `UpdateInterface` is now `SnapInterfaceControllerUpdateInterfaceAction`.
+    - `DeleteInterface` is now `SnapInterfaceControllerDeleteInterfaceAction`.
+    - `UpdateInterfaceState` is now `SnapInterfaceControllerUpdateInterfaceStateAction`.
+    - `ResolveInterface` is now `SnapInterfaceControllerResolveInterfaceAction`.
 - **BREAKING:** All `SnapController` event types were renamed from `OnSomething` to `SnapControllerOnSomethingEvent` ([#3907](https://github.com/MetaMask/snaps/pull/3907))
   - `SnapStateChange` was removed in favour of `SnapControllerStateChangeEvent`.
   - `SnapBlocked` is now `SnapControllerSnapBlockedEvent`.